requirements.txt
epam.indigo